Create builder module
