+++ pyqir/src/python.rs
@@ -2,8 +2,9 @@
 // Licensed under the MIT License.
 
 use crate::{
+    builder::Builder,
     evaluator::PyNonadaptiveJit,
-    generator::{bitcode_to_ir, ir_to_bitcode, r#const, Attribute, BasicQisBuilder, Builder},
+    generator::{bitcode_to_ir, ir_to_bitcode, r#const, Attribute, BasicQisBuilder},
     instructions::{
         Call, FCmp, FloatPredicate, ICmp, Instruction, IntPredicate, Opcode, Phi, Switch,
     },
